+++ src/Sulu/Bundle/ContentBundle/Resources/config/list-builder/BasePageDocument.xml
@@ -2,7 +2,7 @@
        xmlns:orm="http://schemas.sulu.io/class/doctrine"
        xmlns:list="http://schemas.sulu.io/class/list">
     <properties>
-        <property name="id" visibility="no" filter-type="string" list:translation="public.id" />
+        <property name="id" visibility="yes" filter-type="string" list:translation="public.id" />
         <property name="title" visibility="always" filter-type="string" list:translation="public.title" />
     </properties>
 </class>
